+++ TelltaleEditor/3rdparty/lua502/lstrlib.c
@@ -449,7 +449,7 @@ static const char *lmemfind (const char *s1, size_t l1,
 
 
 static void push_onecapture (MatchState *ms, int i) {
-  int l = ms->capture[i].len;
+  int l = (int)ms->capture[i].len;
   if (l == CAP_UNFINISHED) luaL_error(ms->L, "unfinished capture");
   if (l == CAP_POSITION)
     lua_pushnumber(ms->L, (lua_Number)(ms->capture[i].init - ms->src_init + 1));
@@ -527,7 +527,7 @@ static int gfind_aux (lua_State *L) {
     const char *e;
     ms.level = 0;
     if ((e = match(&ms, src, p)) != NULL) {
-      int newstart = e-s;
+      int newstart = (int)(e-s);
       if (e == src) newstart++;  /* empty match? go at least one position */
       lua_pushnumber(L, (lua_Number)newstart);
       lua_replace(L, lua_upvalueindex(3));
